core/app/requirements.txt
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
click==8.1.7
ecdsa==0.18.0
fastapi==0.103.1
greenlet==2.0.2
h11==0.14.0
httptools==0.6.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
orjson==3.9.7
passlib==1.7.4
psycopg2-binary==2.9.7
pyasn1==0.5.0
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
redis==5.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
